refactor: Update User model references to use Quiz and StudyMaterial
feat: Update activity bar colors and title bar colors

feat: Update BASE_URL to use production URL

refactor: Update Quiz model to use cloudinary v2

feat: Add pre-remove hook to delete image from cloudinary
