+++ KrishnaAcademyAPP/utils/uri.tsx
@@ -1,7 +1,7 @@
 // export const SERVER_URI = "https://bookish-fiesta-5pw7j94v67g3pvpw-4000.app.github.dev";
 // export const SERVER_URI = "https://lrzp6q-4000.csb.app"
-// export const SERVER_URI = "https://lms-app-wqwt.onrender.com"
-export const SERVER_URI = "https://lms-app-1-eoim.onrender.com"
+export const SERVER_URI = "https://lms-app-wqwt.onrender.com"
+// export const SERVER_URI = "https://lms-app-1-eoim.onrender.com"
 
 // export const SERVER_URI = "http://localhost:4000"
   // process.env.SERVER_URI || "https://bookish-fiesta-5pw7j94v67g3pvpw-4000.app.github.dev";
